Make Danger Phabricator-friendly. (#32642)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/32642

This Diff adapts the Dangerfile to be a bit more friendly
for PRs that are exported from Phabricator. Specifically,
I'm disabling the check for Test Plan as those are not exported
externally. The check for Summary and Changelog are instead adapted
to work with the format exported for Phabricator.

Changelog:
[Internal] [Changed] - Make Danger Phabricator-friendly

Reviewed By: yungsters

Differential Revision: D32594622

fbshipit-source-id: e31f29defdd926a267cecc9efa6d63de2e580f43