Move react-shallow-renderer from devDependencies to dependencies
Summary:
React Native includes a `ReactNativeTestTools` file that imports `react-shallow-renderer`. This file is used by Jest tests. Since this file might be consumed by tests outside of the RN repo, the dependency should be promoted to `dependencies` to ensure it is actually present when RN is installed from npm.

Changelog: [Internal]

Reviewed By: arushikesarwani94

Differential Revision: D33475662

fbshipit-source-id: 8068294a8b6cade17754699c49a2dfe6d45d86de